// react/src/__generated__/ResourceGroupDefaultSessionOptionsPanel_options.graphql.ts
export type SessionAgentSelectionPolicy = "PREFERRED" | "STRICT" | "%future added value";
export type SessionFailurePolicy = "BOOT_ALL" | "STRICT" | "TOLERANT" | "%future added value";
import { FragmentRefs } from "relay-runtime";
export type ResourceGroupDefaultSessionOptionsPanel_options$data = {
  readonly agentSelectionPolicy: SessionAgentSelectionPolicy;
  readonly clusterMode: string;
  readonly defaultFailurePolicy: SessionFailurePolicy;
  readonly defaultKernelExecutionSpec: {
    readonly batchTimeoutSec: number | null | undefined;
    readonly bootstrapScript: string | null | undefined;
    readonly imageId: string | null | undefined;
    readonly resourceOpts: {
      readonly shmem: string | null | undefined;
    } | null | undefined;
    readonly resources: ReadonlyArray<{
      readonly quantity: any;
      readonly resourceType: string;
    }> | null | undefined;
    readonly startsAt: string | null | undefined;
    readonly startupCommand: string | null | undefined;
  } | null | undefined;
  readonly handlerOptions: {
    readonly byHandler: ReadonlyArray<{
      readonly handlerName: string;
      readonly maxRetryCount: number | null | undefined;
      readonly timeoutSec: number | null | undefined;
    }>;
    readonly default: {
      readonly maxRetryCount: number | null | undefined;
      readonly timeoutSec: number | null | undefined;
    };
  };
  readonly isPreemptible: boolean;
  readonly priority: number;
  readonly " $fragmentType": "ResourceGroupDefaultSessionOptionsPanel_options";
};
export type ResourceGroupDefaultSessionOptionsPanel_options$key = {
  readonly " $data"?: ResourceGroupDefaultSessionOptionsPanel_options$data;
  readonly " $fragmentSpreads": FragmentRefs<"ResourceGroupDefaultSessionOptionsPanel_options">;
};